refactor: migrate task_save tests to task-management-tools.test.ts, delete task-tools.test.ts
